Refactor: Move logic to a customHook
